DRF_24/awesome_project/requirements.txt
django
djangorestframework
drf-spectacular
drf-spectacular[swagger-ui]